+++ invokeai/frontend/web/src/features/parameters/components/Parameters/HighResFix/ParamHrfScale.tsx
@@ -1,13 +1,8 @@
 import { RootState } from 'app/store/store';
 import { useAppDispatch, useAppSelector } from 'app/store/storeHooks';
-import IAIInformationalPopover from 'common/components/IAIInformationalPopover/IAIInformationalPopover';
 import IAISlider from 'common/components/IAISlider';
-import IAISwitch from 'common/components/IAISwitch';
-import {
-  setHrfScale,
-  setHrfToggle,
-} from 'features/parameters/store/generationSlice';
-import { useCallback, ChangeEvent } from 'react';
+import { setHrfScale } from 'features/parameters/store/generationSlice';
+import { useCallback } from 'react';
 
 export default function ParamHrfScale() {
   const hrfScale = useAppSelector(
@@ -26,33 +21,19 @@ export default function ParamHrfScale() {
     [dispatch]
   );
 
-  const hrfToggled = useAppSelector(
-    (state: RootState) => state.generation.hrfToggled
-  );
-
-  const handleHrfToggle = (e: ChangeEvent<HTMLInputElement>) =>
-    dispatch(setHrfToggle(e.target.checked));
-
   return (
-    <IAIInformationalPopover feature="hrf" placement="bottom">
-      <IAISlider
-        label="High Resolution Scale"
-        aria-label="High Resolution Scale"
-        min={1}
-        max={5}
-        step={0.1}
-        value={hrfScale}
-        onChange={handleHrfScaleChange}
-        withSliderMarks
-        withInput
-        withReset
-        handleReset={handleHrfScaleReset}
-      />
-      <IAISwitch
-        label="Toggle High Resolution Fix"
-        isChecked={hrfToggled}
-        onChange={handleHrfToggle}
-      />
-    </IAIInformationalPopover>
+    <IAISlider
+      label="High Resolution Scale"
+      aria-label="High Resolution Scale"
+      min={1}
+      max={5}
+      step={0.1}
+      value={hrfScale}
+      onChange={handleHrfScaleChange}
+      withSliderMarks
+      withInput
+      withReset
+      handleReset={handleHrfScaleReset}
+    />
   );
 }
